+++ src/app/components/content-clusters-spaces-list.js
@@ -9,7 +9,6 @@
 import Component from '@ember/component';
 import { observer, computed } from '@ember/object';
 import { isArray } from '@ember/array';
-import { scheduleOnce } from '@ember/runloop';
 import { inject as service } from '@ember/service';
 import { reads, equal } from '@ember/object/computed';
 import addConflictLabels from 'onedata-gui-common/utils/add-conflict-labels';
@@ -177,15 +176,23 @@ export default Component.extend(
       startRevokeSpace(space) {
         return this.get('startRevokeSpace')(space);
       },
+      /**
+       * @param {Object} supportSpaceData
+       * @param {string} supportSpaceData.storageId
+       * @param {string} supportSpaceData.token
+       * @param {number} supportSpaceData.size
+       * @returns {Promise<{Onepanel.Id}>} Object with ID of the supported space
+       */
       submitSupportSpace(supportSpaceData) {
         const globalNotify = this.get('globalNotify');
         return this.supportSpace(supportSpaceData)
-          .then(() => {
+          .then(result => {
             safeExec(this, 'set', 'supportSpaceOpened', false);
-            scheduleOnce('afterRender', () => this.get('updateSpacesData')());
+            this.updateSpacesData();
             globalNotify.info(
               this.t('supportSuccess')
             );
+            return result;
           });
       },
     },